Managementsystem for Individual Coaching is working properly -> still missing: Member can't ask for a new time

+++ database/migrations/2017_01_10_155554_create_individual_coaching_table.php
@@ -18,7 +18,7 @@ class CreateIndividualCoachingTable extends Migration
             $table->string('services');
             $table->date('date');
             $table->time('time');
-            $table->integer('duration');
+            $table->float('duration');
             $table->float('price_incl_discount');
             $table->boolean('trial');
             $table->boolean('paid');
